+++ src/utils/common.js
@@ -1,0 +1,11 @@
+function getRandomArrayElement(element) {
+  return element[Math.floor(Math.random() * element.length)];
+}
+
+function getRandomNumber(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export {getRandomArrayElement, getRandomNumber};
